fix byteorder

+++ pkg/ebpf/packetinfo.go
@@ -2,11 +2,11 @@ package ebpf
 
 type PacketInfo struct {
 	Pkt                []byte
-	PktId              int
+	PktId              uint
 	MonotoricTimestamp uint64
-	Hookpoint          int
+	Hookpoint          uint
 }
 
 func NewPacketInfo(p []byte, pid int, monot uint64, hook int) *PacketInfo {
-	return &PacketInfo{Pkt: p, PktId: pid, MonotoricTimestamp: monot, Hookpoint: hook}
+	return &PacketInfo{Pkt: p, PktId: uint(pid), MonotoricTimestamp: monot, Hookpoint: uint(hook)}
 }
